backend/requirements.dev.txt
Django>=4.2,<4.3
djangorestframework
django-cors-headers
python-dotenv
openpyxl
pandas
XlsxWriter
requests
weasyprint==61.2
pydyf==0.11.0
qrcode
pillow
